refactor: 빌드 dev, prod 분기 (#49)
