+++ app/src/main/java/com/devhjs/ttackjigeum_android/data/repository/DevPriceHistoryRepositoryImpl.kt
@@ -1,11 +1,12 @@
-package com.devhjs.ttackjigeum_android.data.repository.mock
+package com.devhjs.ttackjigeum_android.data.repository
 
+import com.devhjs.ttackjigeum_android.data.repository.mock.MockData
 import com.devhjs.ttackjigeum_android.domain.model.PriceHistory
 import com.devhjs.ttackjigeum_android.domain.repository.PriceHistoryRepository
 
-
-class MockPriceHistoryRepositoryImpl : PriceHistoryRepository {
+class DevPriceHistoryRepositoryImpl : PriceHistoryRepository {
     override suspend fun getPriceHistories(productId: Long): List<PriceHistory> {
+        // Mock data logic from MockPriceHistoryRepositoryImpl
         return MockData.MockPriceHistories[productId] ?: emptyList()
     }
 }
